scripts/data-generators/requirements.txt
pandas
kafka-python
watchdog
